requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.0
requests>=2.25.0
numpy>=1.21.0
flask>=2.0.0
